--- src/app/Test.tsx
@@ -1,3 +1,0 @@
-export default function Test() {
-  return <span>hello</span>;
-}
